Move new-ledger-funding to new-ledger-channel
