+++ packages/wallet/src/redux/protocols/index.ts
@@ -3,7 +3,7 @@ import { ChallengerState } from './dispute/challenger/states';
 
 import { DirectFundingState } from './direct-funding/states';
 import { FundingState } from './funding/states';
-import { NewLedgerFundingState } from './new-ledger-funding/states';
+import { NewLedgerFundingState } from './new-ledger-channel/states';
 import { ResponderState } from './dispute/responder/states';
 import { WithdrawalState } from './withdrawing/states';
 import { ApplicationState } from './application/states';
