+++ packages/wallet/src/redux/protocols/indirect-funding/__tests__/scenarios.ts
@@ -15,7 +15,7 @@ import { preSuccess as existingLedgerPreSuccess } from '../../existing-ledger-fu
 import {
   preSuccessState as newLedgerPreSuccess,
   successTrigger as newLedgerFundingSuccessTrigger,
-} from '../../new-ledger-funding/__tests__';
+} from '../../new-ledger-channel/__tests__';
 
 const processId = 'processId';
