+++ packages/wallet/src/redux/protocols/new-ledger-channel/components/funding-step.tsx
