+++ packages/wallet/src/redux/protocols/new-ledger-channel/states.ts
